Defined interface
